Rename Spring security class and remove util package
