+++ src/main/java/com/alkemy/ong/config/segurity/SecurityConfig.java
@@ -1,0 +1,22 @@
+package com.alkemy.ong.config.segurity;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@EnableWebSecurity
+@Configuration
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+  @Bean
+  @Override
+  public AuthenticationManager authenticationManagerBean() throws Exception {
+    return super.authenticationManagerBean();
+  }
+
+}
