--- src/main/java/com/alkemy/ong/config/segurity/SegurityConfig.java
@@ -1,14 +1,0 @@
-package com.alkemy.ong.config.segurity;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-public class SegurityConfig extends WebSecurityConfigurerAdapter {
-
-        @Bean
-        @Override
-        public AuthenticationManager authenticationManagerBean() throws Exception {
-            return super.authenticationManagerBean();
-        }
-}
